Serialization now working to and from a file. Docking into editorworkspace and instrumentworkspace needs to be implemented. Bound to S for save and L for load

+++ TeslaRift-Unity/Assets/Scripts/UI/VRControls/InstrumentDock.cs
@@ -17,7 +17,7 @@ namespace VRControls
         // Use this for initialization
         public override void Awake()
         {
-            m_id = (int)VRControls.StaticIds.INSTRUMENT_DOCK;
+            m_id = VRControls.StaticIds.INSTRUMENT_DOCK;
 
             base.Awake();
